sim: Fix incorrect handling of double-writes
The simulator was mistakenly doing something close to what circuits do, keeping
a separate log and merging it, but the merging was incorrect.  Instead we now
keep a log L, snapshot it at the beginning of reach rule into l.

The key observation is that none of the premises of read/write[01] depend only
on ℓ: it's either L only or ℓ ++ L.

This speeds up the processor from 1.097s to 0.97s.