coq: Try a different way to generate switches
The key issue with the previous approach is that it generates quadratic mux
chains for control paths.  For example:

  if c1 then r.r0() else if c2 then s.r0() else t.r0()

When compiling, the .r0 wire of t gets muxed three times:

  Mux(c1, l.t.r0, Mux(c2, l.t.r0, Mux(c3, 1, l.t.r0)))

The new scheme generates this instead:

  v = unpack(0);
  if c1 then v := r.rd0() else pass;
  if c2 then v := s.rd0() else pass;
  if c3 then v := t.rd0() else pass

Which, with redundant Mux optimizations, means that t.r0 can be optimized down
to just Mux(c3, 1, l.t.r0).