sim: Make pack a template to make adding overloads in user code easier
An alternative solution would be to put pack() overloads in the global
namespace to allow argument dependent lookups to find them (the structs are in
the global namespace.  yet another option would be to put both the structs and
the overloads in their own user:: namespace (this way we wouldn't have to prefix
them, which could be nice).