Don't include compiler correctness proofs in the OPAM package
This is a workaround for a problem introduced when moving to the newest dune:
cuttlec often wants to run ocamlc or ocamlopt with -package koika, so we need to
build Dune's @install alias.  The files in coq/CompilerCorrectness take a while
to build, and they aren't needed by cuttlec, so this commit removes them from
the OPAM package to make compilation faster.