coq, sim: Allow extfuns to take *this; use that to build a finish() API
All simulator classes now have a metadata field tracking the information about
the current simulation run.  The finished flag is checked at every cycle to
decide whether to exit the simulation; external functions can set it by calling
finish() (to do that, they need an access to the simulator's ‘this’ pointer,
hence the new flag for external functions, which allows them to request it).
The call to finish() also specifies which information should be printed upon
exiting.

Porting existing code should just be a matter of changing ‘sp_ext_fn_names’ to
‘sp_ext_fn_specs’ and passing a struct with ‘efs_name’ set to the previous name
and ‘efs_method’ set to false, as well as adjusting to the following renames:

- ‘sp_extfuns’ is now ‘sp_prelude’
- ‘ext_fn_spec’ is now ‘ext_fn_rtl_spec’
- ‘ef_name’ and ‘ef_internal’ are now ‘efr_name’ and ‘efr_internal’
- ‘empty_ext_fn_names’ and ‘empty_ext_fn_spec’ are now ‘empty_ext_fn_props’

In C++ code, the main change is that ‘init_and_run’ now returns a ‘snapshot_t’,
not a ‘state_t’.