backends/cpp: Register types returned by functions
This is needed to make sure that we properly import Boost in cases like the
following:

  register r : bits 1
  rule rl =
    write0(r, zext<80>(0b1)[0])

Without the added registration, we wouldn't notice that zext produces a wide
number.  We don't need to add registrations to other branches because leaves are
vars, unops, binops, constants, and functions: constants and functions are
already registered by their respective printers, and vars by the value part of
their binds.